Cleaned up this repo to work with updated Dabo
Many of the imports were structured to assume that the IDE was part of
the Dabo repo; these have been updated to import dabo as an external
library. The imports have been sorted, and the code formatted.

Things still don't always work, but the cxn_editor seems OK. The Class
Designer still throws random exceptions (at least on my Mac), so it
still isn't very usable.